# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Black Pear Joggers API",
        "contact": {
            "name": "Paul Evans",
            "url": "https://bpj.org.uk",
            "email": "paul3vanz@gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/awardclaim": {
            "get": {
                "tags": [
                    "AwardClaims"
                ],
                "summary": "Get all award claims",
                "operationId": "787eed9a536fe3b10fac3620898adb35",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "AwardClaims"
                ],
                "summary": "Submit a club standards award claim",
                "operationId": "935091293cd2b0e9c647b08b0357b003",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "gender",
                                    "category",
                                    "award",
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "races"
                                ],
                                "properties": {
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "W"
                                        ]
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "award": {
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "races": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/awardclaim/{id}/{uniqueToken}": {
            "get": {
                "tags": [
                    "AwardClaims"
                ],
                "summary": "Get an award claim by ID and unique token",
                "operationId": "5c157d5984c9bb8e616cac87722495d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "uniqueToken",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/awardclaim/toggleverified/{id}": {
            "post": {
                "tags": [
                    "AwardClaims"
                ],
                "summary": "Toggle whether an award claim is verified",
                "operationId": "02c3c1cdb27547b91feea04a0f544917",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/awardclaim/archive/{id}": {
            "post": {
                "tags": [
                    "AwardClaims"
                ],
                "summary": "Archive an award claim",
                "operationId": "e7ad8557033783dd265805e83a455509",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/awardclaim/delete/{id}": {
            "post": {
                "tags": [
                    "AwardClaims"
                ],
                "summary": "Delete an award claim",
                "operationId": "f7d9de42fb151e52e43fdd6d11e834e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/awardclaim/{id}": {
            "patch": {
                "tags": [
                    "AwardClaims"
                ],
                "summary": "Update an award claim, and optionally its races",
                "operationId": "bdeb0040817cac34c62cc3cc05343bb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/awardclaim/{id}/race": {
            "post": {
                "tags": [
                    "AwardClaims"
                ],
                "summary": "Update a race on an award claim",
                "operationId": "87c8d4099793360167257552a93ba9a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "race"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "race": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/athletes/awards": {
            "get": {
                "tags": [
                    "Awards"
                ],
                "summary": "Get award-winning performances, optionally filtered by athlete and year",
                "operationId": "ea29f44ca3c3e17f156e56adbd338154",
                "parameters": [
                    {
                        "name": "athlete",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}